backend/requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
functions-framework==3.8.2
google-api-core==2.23.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.3
